+++ src/context/RariContext.tsx
@@ -29,6 +29,81 @@ import {
 import { Web3Provider } from "@ethersproject/providers";
 import { ChainID, isSupportedChainId } from "esm/utils/networks";
 
+interface ChainMetadata {
+  chainId: number;
+  name: string;
+  imageUrl?: string;
+  supported: boolean;
+  rpcUrl: string;
+}
+
+const chainMetadata = {
+  [ChainID.ETHEREUM]: {
+    chainId: ChainID.ETHEREUM,
+    name: "Ethereum",
+    imageUrl: "/static/networks/ethereum.png",
+    supported: true,
+    rpcUrl:
+      "https://eth-mainnet.alchemyapi.io/v2/2Mt-6brbJvTA4w9cpiDtnbTo6qOoySnN",
+  },
+  [ChainID.ROPSTEN]: {
+    chainId: ChainID.ROPSTEN,
+    name: "Ropsten",
+    supported: false,
+    rpcUrl: "https://ropsten.infura.io/v3/9aa3d95b3bc440fa88ea12eaa4456161",
+  },
+  [ChainID.RINKEBY]: {
+    chainId: ChainID.RINKEBY,
+    name: "Rinkeby",
+    supported: false,
+    rpcUrl: "https://rinkeby.infura.io/v3/9aa3d95b3bc440fa88ea12eaa4456161",
+  },
+  [ChainID.GÖRLI]: {
+    chainId: ChainID.GÖRLI,
+    name: "Goerli",
+    supported: false,
+    rpcUrl: "https://goerli.infura.io/v3/9aa3d95b3bc440fa88ea12eaa4456161",
+  },
+  [ChainID.KOVAN]: {
+    chainId: ChainID.KOVAN,
+    name: "Kovan",
+    supported: false,
+    rpcUrl: "https://kovan.infura.io/v3/9aa3d95b3bc440fa88ea12eaa4456161",
+  },
+  [ChainID.ARBITRUM]: {
+    chainId: ChainID.ARBITRUM,
+    name: "Arbitrum",
+    imageUrl: "/static/networks/arbitrum.svg",
+    supported: false,
+    rpcUrl: "https://arb1.arbitrum.io/rpc",
+  },
+  [ChainID.ARBITRUM_TESTNET]: {
+    chainId: ChainID.ARBITRUM_TESTNET,
+    name: "Arbitrum Testnet",
+    imageUrl: "/static/networks/arbitrum.svg",
+    supported: true,
+    rpcUrl:
+      "https://arb-rinkeby.g.alchemy.com/v2/PkZ7ilUhTBT6tHUsgToel62IOcuyKcwb",
+  },
+  [ChainID.OPTIMISM]: {
+    chainId: ChainID.OPTIMISM,
+    name: "Optimism",
+    imageUrl: "/static/networks/optimism.svg",
+    supported: false,
+    rpcUrl: "https://mainnet.optimism.io/",
+  },
+};
+
+export function getSupportedChains(): ChainMetadata[] {
+  return Object.values(chainMetadata).filter(
+    (chainMetadata) => chainMetadata.supported
+  );
+}
+
+export function getChainMetadata(chainId: ChainID): ChainMetadata {
+  return chainMetadata[chainId];
+}
+
 async function launchModalLazy(
   t: (text: string, extra?: any) => string,
   cacheProvider: boolean = true
@@ -109,6 +184,7 @@ export interface RariContextData {
   address: string;
   isAttemptingLogin: boolean;
   chainId: number | undefined;
+  switchNetwork: (newChainId: number) => void;
 }
 
 export const EmptyAddress = "0x0000000000000000000000000000000000000000";
@@ -177,7 +253,8 @@ export const RariProvider = ({ children }: { children: ReactNode }) => {
 
       console.log({ chainId });
 
-      let _chainId = chainId === ChainID.ETHEREUM_TESTNET ? ChainID.ETHEREUM : chainId
+      let _chainId =
+        chainId === ChainID.ETHEREUM_TESTNET ? ChainID.ETHEREUM : chainId;
 
       const rariInstance = new Vaults(provider);
       const fuseInstance = initFuseWithProviders(provider, _chainId);
@@ -260,6 +337,42 @@ export const RariProvider = ({ children }: { children: ReactNode }) => {
     };
   }, [web3ModalProvider, refetchAccountData]);
 
+  // Based on Metamask-recommended code at
+  // https://docs.metamask.io/guide/rpc-api.html#usage-with-wallet-switchethereumchain
+  // TODO(nathanhleung) handle all possible errors
+  const switchNetwork = async function (chainId: ChainID) {
+    const hexChainId = chainId.toString(16);
+    const chainMetadata = getChainMetadata(chainId);
+
+    try {
+      await window.ethereum.request({
+        method: "wallet_switchEthereumChain",
+        params: [{ chainId: `0x${hexChainId}` }],
+      });
+    } catch (switchError) {
+      // This error code indicates that the chain has not been added to MetaMask.
+      if ((switchError as any).code === 4902) {
+        try {
+          await window.ethereum.request({
+            method: "wallet_addEthereumChain",
+            params: [
+              {
+                chainId: `0x${hexChainId}`,
+                chainName: chainMetadata.name,
+                rpcUrls: [chainMetadata.rpcUrl],
+              },
+            ],
+          });
+        } catch (addError) {
+          // handle "add" error
+        }
+      }
+      // handle other "switch" errors
+    } finally {
+      refetchAccountData();
+    }
+  };
+
   const value = useMemo(
     () => ({
       web3ModalProvider,
@@ -271,6 +384,7 @@ export const RariProvider = ({ children }: { children: ReactNode }) => {
       address,
       isAttemptingLogin,
       chainId,
+      switchNetwork,
     }),
     [rari, web3ModalProvider, login, logout, address, fuse, isAttemptingLogin]
   );
